+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # v1.9.3
+## Updates
+* **Status:** Added `x/month` & `x/year` to P2Pool stats
+
 ## Fixes
 * **Wallet:** List will wrap to stay within `5` wallets or `40` characters per line
 * **Status:** Fixed P2Pool `Latest share` showing 1st share found